Project Completed

+++ src/main/java/com/ashu/investment_calculator/model/InvestmentDetails.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class InvestmentDetails {
     private int year;
-    private int initialInvestment;
-    private int yearlyInvestment;
-    private int totalInvestment;
-    private int interestEarn;
-    private int totalInterest;
-    private int endingBalance;
+    private double initialInvestment;
+    private double yearlyInvestment;
+    private double totalInvestment;
+    private double interestEarn;
+    private double totalInterest;
+    private double endingBalance;
 
 }
